requirements.txt
anyio==4.4.0
backoff==2.2.1
certifi==2024.8.30
charset-normalizer==3.3.2
exceptiongroup==1.2.2
gql==3.5.0
graphql-core==3.2.4
idna==3.9
multidict==6.1.0
python-dotenv==1.0.1
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.11.1
